Move position details into its own component, externalize detailsObject to be used across tests
